requirements.txt
typer
boto3
rich
PyYAML
questionary